Refactoring the source path and incorporating the suggestions in #8
