--- StardewCropCalc/com.elizabethburgessamt.stardewcropcalc/month/Month.java
@@ -1,7 +1,0 @@
-package month;
-
-public enum Month {
-	SPRING,
-	SUMMER,
-	FALL
-}
