+++ StardewCropCalc/src/com/elizabethburgessamt/stardewcropcalc/fields/Field.java
@@ -1,0 +1,8 @@
+package com.elizabethburgessamt.stardewcropcalc.fields;
+
+public interface Field {
+	
+	//static int cropCapacity;
+	//static int trellisCapacity;
+
+}
