+++ StardewCropCalc/src/com/elizabethburgessamt/stardewcropcalc/fields/ThirteenSquare.java
@@ -1,0 +1,5 @@
+package com.elizabethburgessamt.stardewcropcalc.fields;
+
+public class ThirteenSquare implements Field {
+
+}
