+++ StardewCropCalc/src/com/elizabethburgessamt/stardewcropcalc/month/Month.java
@@ -1,0 +1,7 @@
+package com.elizabethburgessamt.stardewcropcalc.month;
+
+public enum Month {
+	SPRING,
+	SUMMER,
+	FALL
+}
