+++ StardewCropCalc/src/com/elizabethburgessamt/stardewcropcalc/seeds/SeedMaker.java
@@ -1,4 +1,4 @@
-package seeds;
+package com.elizabethburgessamt.stardewcropcalc.seeds;
 
 public class SeedMaker {
 	//TODO: File should read in and create seeds based on input month.
